pyproject.toml
[tool.poetry]
name = "metamask"
version = "0.1.0"
description = "Metamask bot for making transactions"
authors = ["sa1am8 <tinkerino571@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
web3 = "^6.15.1"
eth-account = "^0.11.0"
pydantic-settings = "^2.2.0"
websocket-client = "^1.7.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
ruff = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
